Refactor UPC track skimmer and candidate producer (#1334)
