+++ PWGUD/TableProducer/CMakeLists.txt
@@ -19,11 +19,6 @@ o2physics_add_dpl_workflow(dgbccand-producer
                            PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore O2Physics::DGCutparHolder
                            COMPONENT_NAME Analysis)
 
-o2physics_add_dpl_workflow(upctrack-skimmer
-                           SOURCES UPCTrackSkimmer.cxx
-                           PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore
-                           COMPONENT_NAME Analysis)
-
 o2physics_add_dpl_workflow(upccand-producer
                            SOURCES UPCandidateProducer.cxx
                            PUBLIC_LINK_LIBRARIES O2::Framework O2Physics::AnalysisCore O2Physics::AnalysisCCDB
